requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg==3.2.4
psycopg-binary==3.2.4
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pytest==8.3.5
pytest-asyncio==0.25.3
python-dotenv==1.0.1
setuptools==75.8.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
wheel==0.45.1
